[Refactor]-Product 관련 기능 코드 개선, api 파일 분리, 전역 상태 제거
