+++ src/lib/product/hooks/useDeleteProduct.ts
@@ -1,11 +1,9 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { doc, deleteDoc } from "firebase/firestore";
 import { db } from "@/firebase";
-import { useProductStore } from "@/store/product/useProductStore";
 
 export const useDeleteProduct = () => {
   const queryClient = useQueryClient();
-  const deleteProduct = useProductStore((state) => state.deleteProduct);
 
   return useMutation({
     mutationFn: async (productId: string) => {
@@ -13,12 +11,13 @@ export const useDeleteProduct = () => {
       await deleteDoc(productRef);
       return productId;
     },
-    onSuccess: (productId) => {
+    onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["products"] });
-      deleteProduct(productId);
+      // toast
     },
     onError: (error) => {
       console.error("상품 삭제 중 에러 발생:", error);
+      // toast
     },
   });
 };
